fix: 게임이 끝나지 않아도 데이터가 지워지는 오류 수정

+++ src/main/java/chess/service/ChessGameService.java
@@ -32,8 +32,11 @@ public class ChessGameService {
     }
 
     public boolean gameOver() {
-        chessDao.deleteNotation();
-        return chessGame.isGameOver();
+        if (chessGame.isGameOver()) {
+            chessDao.deleteNotation();
+            return true;
+        }
+        return false;
     }
 
     public Score getScore() {
